[DS-3158] Move the command method somewhere less deprecated.
This code no longer has any connection with the deprecated class
Configurationmanager, so move it now before we lose it by accident.
